luci/requirements/common.txt
discord==1.0.1

python-decouple==3.1
requests==2.22.0
flask==1.1.2
